// src/models/MediaInfo.ts
export type MediaInfo = {
  title: string;
  name: string;
  original_name: string;
  original_title: string;
  backdrop_path: string;
  overview: string;
  vote_average: number;
  vote_count: number;
  release_date: string;
  runtime: number;
  id: number;
  media_type: string;
  adult: boolean;
  poster_path: string;
  gennre_ids: number[];
  popularity: number;
  video: boolean;
  original_language: string;
  number_of_episodes: number;
  number_of_seasons: number;
  genres: { name: string; id: number }[];
  first_air_date: string;
};